server/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-dotenv==1.1.0
twilio==9.5.0
pydantic==2.11.0
